cmd/cueckoo: add unity command
Includes a refactor of common CL-based trigger code into a type shared
between unity and runtrybot.
