+++ cmd/cueckoo/cmd/testdata/unity_cl.golden
@@ -1,0 +1,14 @@
+{
+  "event_type": "unity run for ref",
+  "client_payload": {
+    "type": "unity",
+    "payload": {
+      "cl": {
+        "changeID": "change",
+        "ref": "ref",
+        "commit": "commit"
+      },
+      "versions": "\"ref\""
+    }
+  }
+}
