+++ cmd/cueckoo/cmd/testdata/unity_versions.golden
@@ -1,0 +1,10 @@
+{
+  "event_type": "hello",
+  "client_payload": {
+    "type": "unity",
+    "payload": {
+      "cl": null,
+      "versions": "\"v0.3.0-beta.5\""
+    }
+  }
+}
